Move loi converter
